requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
filelock==3.0.12
httplib2==0.14.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.15.2
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.5.0
macaroonbakery==1.3.1
MarkupSafe==1.1.0
more-itertools==8.2.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==20.3
parsel==1.5.2
parso==0.5.2
pathlib==1.0.1
pdf2image==1.13.1
pdfminer.six==20200517
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prompt-toolkit==2.0.10
Protego==0.1.16
protobuf==3.6.1
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycryptodome==3.9.7
pycups==1.9.73
PyDispatcher==2.0.5
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
pytesseract==0.3.4
pytest==5.4.2
python-apt==2.0.0
python-debian===0.1.36ubuntu1
pytz==2019.3
PyYAML==5.3.1
queuelib==1.5.0
requests==2.23.0
requests-unixsocket==0.2.0
Scrapy==2.1.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sortedcontainers==2.1.0
ssh-import-id==5.10
systemd-python==234
traitlets==4.3.3
Twisted==20.3.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
virtualenv==20.0.21
w3lib==1.21.0
wadllib==1.3.3
wcwidth==0.1.9
zipp==1.0.0
zope.interface==5.1.0
